+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [1.63.2] - 2019-11-XX
-Fix comment fo admin::createCompagny, Country is mandatory.
+-   Fix comment fo admin::createCompagny, Country is mandatory.
+-   Add method bubbles::inviteContactsByEmailsToBubble to Invite a list of contacts by emails in a bubble.
+
 
 ## [1.63.1] - 2019-11-06
 -   Add start duration in result of start.